+++ odoo/local-src/specific_partner/models/res_partner.py
@@ -3,6 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import api, fields, models, _
+from odoo.exceptions import ValidationError
 
 
 class ResPartner(models.Model):
@@ -57,6 +58,27 @@ class ResPartner(models.Model):
             partner.is_veterinary = (
                 partner.alcyon_category_id == veterinary
             )
+
+    @api.one
+    @api.constrains('name', 'street', 'city', 'zip', 'country_id')
+    def _is_valid_esb_address(self):
+        """Check customer address validity.
+
+        Invoicing and delivery address of customer are sent to the esb,
+        and some fields are required for them to be valid.
+        And check is made in the view as well.
+        """
+        if not self.parent_id or not self.customer:
+            return
+        if self.type not in ['invoice', 'delivery']:
+            return
+        if (self.name and self.street and self.city and
+                self.zip and self.country_id.esb_ref):
+            return
+        raise ValidationError(_('For an invoicing or delivery address the '
+                                'following fields (name, street, city, zip, '
+                                'country) are required. And the country '
+                                'must have a reference ESB.'))
 
     _sql_constraints = [
             ('ref_digit_only',