+++ HISTORY.rst
@@ -18,6 +18,7 @@ Unreleased
 * ALCN-1415: Improve esb export so records being updated while an export is running will not be missed on next export.
              The scheduled exports to the esb are run with queue_job.
 * ALCN-1442: Sale order in draft state are not to be exported to esb/magento.
+* ALCN-1440: Add some required fields for customer invoicing and delivery addresses, so they can be sent to the ESB.
 
 **Bugfixes**